backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bitarray==3.8.0
cached-property==2.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
cytoolz==1.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
mangum==0.21.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.4.1
packaging==26.0
pandas==3.0.0
parsimonious==0.10.0
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
rlp==4.1.0
scikit-learn==1.8.0
scipy==1.17.0
sentry-sdk==2.50.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlalchemy-orm==1.2.10
starlette==0.50.0
tenacity==9.1.3
threadpoolctl==3.6.0
toolz==1.1.0
tqdm==4.67.3
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
